requirements_hf.txt
websockets==12.0
transformers
torch
torchaudio
opencc-python-reimplemented
accelerate
